Minor fixes in various files
- Change the argument types of the color class methods from char to int
- Replace fmod() function with fmodf()
- Fix some warnings with return values in the GL hashmap class
- Set the library name debug postfix to nothing on all platforms
- Change the default message for asserts without an user message
- Fix an assert in the animated sprite class
- Set the external Android directory variable regardless of the build flag
- Add the /MP option to compile with multiple processes on MSVC
- Add the /fp:fast option to MSVC for faster floating-point operations

+++ include/SceneNode.h
@@ -156,7 +156,7 @@ class DLL_PUBLIC SceneNode : public Object
 
 inline void SceneNode::setRotation(float rotation)
 {
-	rotation_ = fmod(rotation, 360.0f);
+	rotation_ = fmodf(rotation, 360.0f);
 }
 
 }
